chore(example): Clean up Vite example
