+++ apps/vite-vue3/.eslintrc.cjs
@@ -9,7 +9,4 @@ module.exports = {
     "@vue/eslint-config-typescript/recommended",
     "@vue/eslint-config-prettier",
   ],
-  env: {
-    "vue/setup-compiler-macros": true,
-  },
 };
