+++ apps/vite-vue3/env.d.ts
