+++ apps/vite-vue3/index.html
@@ -2,9 +2,8 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
+    <title>tsParticles Vite + Vue 3 Demo</title>
   </head>
   <body>
     <div id="app"></div>
