+++ apps/vite-vue3/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vue3-demo",
+  "name": "vite-vue3-demo",
   "version": "2.12.0",
   "private": true,
   "scripts": {
@@ -17,14 +17,13 @@
     "tsparticles-engine": "^2.12.0",
     "vue": "^3.3.4",
     "vue-router": "^4.2.4",
-    "vue3-particles": "workspace:^"
+    "@tsparticles/vue": "workspace:^"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.3.2",
     "@tsconfig/node18": "^18.2.0",
     "@types/node": "^20.4.7",
     "@vitejs/plugin-vue": "^4.2.3",
-    "@vitejs/plugin-vue-jsx": "^3.0.1",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/eslint-config-typescript": "^11.0.3",
     "@vue/tsconfig": "^0.4.0",
