--- apps/vue3/src/App.vue
@@ -1,89 +1,0 @@
-<script setup lang="ts">
-import type { Container, Engine } from "tsparticles-engine";
-import { loadFull } from "tsparticles";
-import configs from "tsparticles-demo-configs";
-
-const options = configs.basic;
-
-const particlesInit = async (engine: Engine) => {
-  await loadFull(engine);
-};
-
-const particlesLoaded = async (container: Container) => {
-  console.log(container);
-};
-</script>
-
-<template>
-  <main>
-    <vue-particles
-      id="tsparticles"
-      :options="options"
-      :particles-init="particlesInit"
-      @particles-loaded="particlesLoaded"
-    />
-  </main>
-</template>
-
-<style>
-@import "./assets/base.css";
-
-#app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-
-  font-weight: normal;
-}
-
-header {
-  line-height: 1.5;
-}
-
-.logo {
-  display: block;
-  margin: 0 auto 2rem;
-}
-
-a,
-.green {
-  text-decoration: none;
-  color: hsla(160, 100%, 37%, 1);
-  transition: 0.4s;
-}
-
-@media (hover: hover) {
-  a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    padding: 0 2rem;
-  }
-
-  header {
-    display: flex;
-    place-items: center;
-    padding-right: calc(var(--section-gap) / 2);
-  }
-
-  header .wrapper {
-    display: flex;
-    place-items: flex-start;
-    flex-wrap: wrap;
-  }
-
-  .logo {
-    margin: 0 2rem 0 0;
-  }
-}
-</style>
